requirements.txt
Package                Version
---------------------- ---------
et_xmlfile             2.0.0
mysql-connector-python 9.4.0
numpy                  2.0.2
opencv-contrib-python  4.13.0.92
openpyxl               3.1.5
Pillow                 8.1.0
pip                    26.0.1
setuptools             49.2.1
cd



##  Note: Python version -----> Python 3.9.1